Add files via upload
Here it is a imageclassification model created using Machine Learning technology . This uses Sequential keras model and compare the input image with the trianing dataset and giving the output it as Dog or a Cat. The model's accuracy is 85% and I hope that it will increased after exploring the model after some time.